% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.R
\name{chi_age}
\alias{chi_age}
\title{Determine an age from Community Health Index (CHI) number}
\usage{
chi_age(.data, ref_date, cutoff_2000 = 20)
}
\arguments{
\item{.data}{Character. A vector of CHIs as characters/strings.}

\item{ref_date}{Dates. Single date or vector with same length as \code{.data}.}

\item{cutoff_2000}{Integer. Two-digit year before which is considered 2000s.}
}
\value{
A vector of ages.
}
\description{
Determine an age from Community Health Index (CHI) number
}
\examples{
library(dplyr)
# These CHIs are not real.
chi = c("1009701234", "1811431232", "1304496368")

# Age today
chi_age(chi, Sys.time())

# Age on a single date
library(lubridate)
chi_age(chi, dmy("11/09/2018"))

# Age on a vector of dates
dates = dmy("11/09/2018",
            "09/05/2015",
            "10/03/2014")
chi_age(chi, dates)
# From tibble
tibble(chi = chi) \%>\%
  mutate(
    age = chi_age(chi, Sys.time())
  )
}
