% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.R
\name{chi_gender}
\alias{chi_gender}
\title{Extract gender from Community Health Index (CHI) number}
\usage{
chi_gender(.data)
}
\arguments{
\item{.data}{Character. A vector of CHIs as characters/strings.}
}
\value{
Factor with two levels "Female", "Male"
}
\description{
Extract gender from Community Health Index (CHI) number
}
\examples{
library(dplyr)
# These CHIs are not real.
chi = c("1009701234", "1811431232", "1304496368")

chi_gender(chi)
# From tibble
tibble(chi = chi) \%>\%
  mutate(
    gender = chi_gender(chi)
  )
}
