% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.R
\name{chi_valid}
\alias{chi_valid}
\title{Test for valid Community Health Index (CHI) number}
\usage{
chi_valid(.data)
}
\arguments{
\item{.data}{Character. A vector of CHIs as characters/strings.}
}
\value{
A logical vector with \code{FALSE} indicating a non-valid CHI.
}
\description{
Test for valid Community Health Index (CHI) number
}
\examples{
library(dplyr)
# These CHIs are not real.
chi = c("1009701234", "1811431232", "1304496368")

chi_valid(chi)
# From tibble
tibble(chi = chi) \%>\%
  mutate(
    chi_valid = chi_valid(chi)
  )
}
