% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.R
\name{clean_chi}
\alias{clean_chi}
\title{Clean CHI}
\usage{
clean_chi(.data)
}
\arguments{
\item{.data}{}
}
\description{
Clean CHI
}
\examples{
chi = c("1009701234", "1811431232", "1304496368",
  "10 10 19 1234", "   12 12 30 1 2 3 4    ")
  clean_chi(chi)

# Extra digit will error
chi = "1009701234 3"
}
\keyword{internal}
